Simple fix for diamond shop songs.
Not the best fix but it works.
Cancion Por Ferrer, Doki Doki Summer Days, Must Be Nice, and Wishin' I Had It Now were already unlocked, originally

The rest of them, I took the internal names of them from the unlock list included with the game, added those to a new list in the source file, and cloned the entry for Cancion Por Ferrer and replaced the name with the internal names.

It works, but it's not perfect. I assume that "DT_UnlocksSongsData.h" is a file that has the original unlocks data before the custom songs get added. Better fix would be to update that with the latest one, but I'm not sure how that is formatted exactly so this works for now.